requirements.txt
pyTelegramBotAPI 4.28.0
python-dotenv 1.1.1
requests 2.32.4